{"schema":"decodex/commit/1","summary":"Modularize selection size badge affordances","authority":"manual"}
